Split ProjectDependencyGraph.cs into parts for high-level operations
