+++ src/Workspaces/Core/Portable/Workspace/Solution/ProjectDependencyGraph.cs
@@ -5,8 +5,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using Roslyn.Utilities;
@@ -16,7 +14,7 @@ namespace Microsoft.CodeAnalysis
     /// <summary>
     /// A <see cref="ProjectDependencyGraph"/> models the dependencies between projects in a solution.
     /// </summary>
-    public class ProjectDependencyGraph
+    public partial class ProjectDependencyGraph
     {
         private readonly ImmutableHashSet<ProjectId> _projectIds;
         private readonly ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> _referencesMap;
@@ -80,487 +78,6 @@ namespace Microsoft.CodeAnalysis
             _reverseTransitiveReferencesMap = reverseTransitiveReferencesMap;
             _lazyTopologicallySortedProjects = topologicallySortedProjects;
             _lazyDependencySets = dependencySets;
-        }
-
-        internal ProjectDependencyGraph WithAdditionalProjects(IEnumerable<ProjectId> projectIds)
-        {
-            // Track the existence of some new projects. Note this call only adds new ProjectIds, but doesn't add any references. Any caller who wants to add a new project
-            // with references will first call this, and then call WithAdditionalProjectReferences to add references as well.
-
-            // Since we're adding a new project here, there aren't any references to it, or at least not yet. (If there are, they'll be added
-            // later with WithAdditionalProjectReferences). Thus, the new projects aren't topologically sorted relative to any other project
-            // and form their own dependency set. Thus, sticking them at the end is fine.
-            var newTopologicallySortedProjects = _lazyTopologicallySortedProjects;
-
-            if (!newTopologicallySortedProjects.IsDefault)
-            {
-                newTopologicallySortedProjects = newTopologicallySortedProjects.AddRange(projectIds);
-            }
-
-            var newDependencySets = _lazyDependencySets;
-
-            if (!newDependencySets.IsDefault)
-            {
-                var builder = newDependencySets.ToBuilder();
-
-                foreach (var projectId in projectIds)
-                {
-                    builder.Add(ImmutableArray.Create(projectId));
-                }
-
-                newDependencySets = builder.ToImmutable();
-            }
-
-            // The rest of the references map is unchanged, since no new references are added in this call.
-            return new ProjectDependencyGraph(
-                _projectIds.Union(projectIds),
-                referencesMap: _referencesMap,
-                reverseReferencesMap: _lazyReverseReferencesMap,
-                transitiveReferencesMap: _transitiveReferencesMap,
-                reverseTransitiveReferencesMap: _reverseTransitiveReferencesMap,
-                topologicallySortedProjects: newTopologicallySortedProjects,
-                dependencySets: newDependencySets);
-        }
-
-        internal ProjectDependencyGraph WithAdditionalProjectReferences(ProjectId projectId, IReadOnlyList<ProjectId> referencedProjectIds)
-        {
-            Contract.ThrowIfFalse(_projectIds.Contains(projectId));
-
-            if (referencedProjectIds.Count == 0)
-            {
-                return this;
-            }
-
-            var newReferencesMap = ComputeNewReferencesMapForAdditionalProjectReferences(_referencesMap, projectId, referencedProjectIds);
-
-            var newReverseReferencesMap =
-                _lazyReverseReferencesMap != null
-                    ? ComputeNewReverseReferencesMapForAdditionalProjectReferences(_lazyReverseReferencesMap, projectId, referencedProjectIds)
-                    : null;
-
-            var newTransitiveReferencesMap = ComputeNewTransitiveReferencesMapForAdditionalProjectReferences(_transitiveReferencesMap, projectId, referencedProjectIds);
-
-            var newReverseTransitiveReferencesMap = ComputeNewReverseTransitiveReferencesMapForAdditionalProjectReferences(_reverseTransitiveReferencesMap, projectId, referencedProjectIds);
-
-            // Note: rather than updating our dependency sets and topologically sorted data, we'll throw that away since incremental update is
-            // tricky, and those are rarely used. If somebody needs them, it'll be lazily computed.
-            return new ProjectDependencyGraph(
-                _projectIds,
-                referencesMap: newReferencesMap,
-                reverseReferencesMap: newReverseReferencesMap,
-                transitiveReferencesMap: newTransitiveReferencesMap,
-                reverseTransitiveReferencesMap: newReverseTransitiveReferencesMap,
-                topologicallySortedProjects: default,
-                dependencySets: default);
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_referencesMap"/> for the addition of additional project references.
-        /// </summary>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReferencesMapForAdditionalProjectReferences(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReferencesMap,
-            ProjectId projectId,
-            IReadOnlyList<ProjectId> referencedProjectIds)
-        {
-            if (existingReferencesMap.TryGetValue(projectId, out var existingReferences))
-            {
-                return existingReferencesMap.SetItem(projectId, existingReferences.Union(referencedProjectIds));
-            }
-            else
-            {
-                return existingReferencesMap.SetItem(projectId, referencedProjectIds.ToImmutableHashSet());
-            }
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_lazyReverseReferencesMap"/> for the addition of additional project references.
-        /// Must be called on a non-null map.
-        /// </summary>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseReferencesMapForAdditionalProjectReferences(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseReferencesMap,
-            ProjectId projectId,
-            IReadOnlyList<ProjectId> referencedProjectIds)
-        {
-            var builder = existingReverseReferencesMap.ToBuilder();
-
-            foreach (var referencedProject in referencedProjectIds)
-            {
-                if (builder.TryGetValue(referencedProject, out var reverseReferences))
-                {
-                    builder[referencedProject] = reverseReferences.Add(projectId);
-                }
-                else
-                {
-                    builder[referencedProject] = ImmutableHashSet.Create(projectId);
-                }
-            }
-
-            return builder.ToImmutable();
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_lazyReverseReferencesMap"/> for the removal of a project.
-        /// </summary>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseReferencesMapForRemovedProject(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingForwardReferencesMap,
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseReferencesMap,
-            ProjectId projectId)
-        {
-            var builder = existingReverseReferencesMap.ToBuilder();
-
-            // Iterate over each project referenced by 'projectId', which is now being removed. Update the reverse
-            // references map for the project to no longer include 'projectId' in the list.
-            if (existingForwardReferencesMap.TryGetValue(projectId, out var referencedProjectIds))
-            {
-                foreach (var referencedProjectId in referencedProjectIds)
-                {
-                    if (builder.TryGetValue(referencedProjectId, out var reverseDependencies))
-                    {
-                        builder[referencedProjectId] = reverseDependencies.Remove(projectId);
-                    }
-                }
-            }
-
-            // Finally, remove 'projectId' itself.
-            builder.Remove(projectId);
-            return builder.ToImmutable();
-        }
-
-        private ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseReferencesMapForRemovedProjectReference(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseReferencesMap,
-            ProjectId projectId,
-            ProjectId referencedProjectId)
-        {
-            if (existingReverseReferencesMap.TryGetValue(referencedProjectId, out var referencingProjects))
-            {
-                return existingReverseReferencesMap.SetItem(referencedProjectId, referencingProjects.Remove(projectId));
-            }
-
-            return existingReverseReferencesMap;
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_referencesMap"/> for the removal of a project.
-        /// </summary>
-        /// <param name="existingReverseReferencesMap">The <see cref="_lazyReverseReferencesMap"/> prior to the removal.
-        /// This map serves as a hint to the removal process; i.e. it is assumed correct if it contains data, but may be
-        /// omitted without impacting correctness.</param>
-        /// <param name="referencesMap">The <see cref="_referencesMap"/> prior to the removal.</param>
-        /// <param name="projectId">The ID of the project which is being removed.</param>
-        /// <returns>The <see cref="_referencesMap"/> for the project dependency graph once the project is removed.</returns>
-        private ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReferencesMapForRemovedProject(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>>? existingReverseReferencesMap,
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> referencesMap,
-            ProjectId projectId)
-        {
-            var builder = referencesMap.ToBuilder();
-
-            if (existingReverseReferencesMap is object && existingReverseReferencesMap.TryGetValue(projectId, out var referencingProjects))
-            {
-                // We know all the projects directly referencing 'projectId', so remove 'projectId' from the set of
-                // references in each of those cases directly.
-                foreach (var id in referencingProjects)
-                {
-                    builder[id] = builder[id].Remove(projectId);
-                }
-            }
-            else
-            {
-                // We don't know which projects reference 'projectId', so iterate over all known projects and remove
-                // 'projectId' from the set of references if it exists.
-                foreach (var (id, references) in referencesMap)
-                {
-                    builder[id] = references.Remove(projectId);
-                }
-            }
-
-            // Finally, remove 'projectId' itself.
-            builder.Remove(projectId);
-            return builder.ToImmutable();
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_transitiveReferencesMap"/> for the removal of a project.
-        /// </summary>
-        /// <seealso cref="ComputeNewReverseTransitiveReferencesMapForRemovedProject"/>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewTransitiveReferencesMapForRemovedProject(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingTransitiveReferencesMap,
-            ProjectId projectId)
-        {
-            var builder = existingTransitiveReferencesMap.ToBuilder();
-
-            // Iterate over each project and invalidate the transitive references for the project if the project has an
-            // existing transitive reference to 'projectId'.
-            foreach (var (project, references) in existingTransitiveReferencesMap)
-            {
-                if (references.Contains(projectId))
-                {
-                    // The project transitively referenced 'projectId', so any transitive references brought in
-                    // exclusively through this reference are no longer valid. Remove the project from the map and the
-                    // new transitive references will be recomputed the first time they are needed.
-                    builder.Remove(project);
-                }
-            }
-
-            // Finally, remove 'projectId' itself.
-            builder.Remove(projectId);
-            return builder.ToImmutable();
-        }
-
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewTransitiveReferencesMapForRemovedProjectReference(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingTransitiveReferencesMap,
-            ProjectId projectId,
-            ProjectId referencedProjectId)
-        {
-            var builder = existingTransitiveReferencesMap.ToBuilder();
-
-            // Invalidate the transitive references from every project referencing the changed project (transitively)
-            foreach (var (project, references) in existingTransitiveReferencesMap)
-            {
-                if (references.Contains(projectId))
-                {
-                    builder.Remove(project);
-                }
-            }
-
-            // Invalidate the transitive references from the changed project
-            builder.Remove(projectId);
-
-            return builder.ToImmutable();
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_transitiveReferencesMap"/> for the addition of additional project references. 
-        /// </summary>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewTransitiveReferencesMapForAdditionalProjectReferences(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingTransitiveReferencesMap,
-            ProjectId projectId,
-            IReadOnlyList<ProjectId> referencedProjectIds)
-        {
-            // To update our forward transitive map, we need to add referencedProjectIds (and their transitive dependencies) to the transitive references
-            // of projects. First, let's just compute the new set of transitive references. It's possible while doing so we'll discover that we don't
-            // know the transitive project references for one of our new references. In that case, we'll use null as a sentinel to mean "we don't know" and
-            // we propagate the not-knowingness. But let's not worry about that yet. First, let's just get the new transitive reference set.
-            HashSet<ProjectId>? newTransitiveReferences = new HashSet<ProjectId>(referencedProjectIds);
-
-            foreach (var referencedProjectId in referencedProjectIds)
-            {
-                if (existingTransitiveReferencesMap.TryGetValue(referencedProjectId, out var additionalTransitiveReferences))
-                {
-                    newTransitiveReferences.UnionWith(additionalTransitiveReferences);
-                }
-                else
-                {
-                    newTransitiveReferences = null;
-                    break;
-                }
-            }
-
-            // We'll now loop through each entry in our existing cache and compute updates. We'll accumulate them into this builder.
-            var builder = existingTransitiveReferencesMap.ToBuilder();
-
-            foreach (var projectIdToUpdate in existingTransitiveReferencesMap.Keys)
-            {
-                existingTransitiveReferencesMap.TryGetValue(projectIdToUpdate, out var existingTransitiveReferences);
-
-                // The projects who need to have their caches updated are projectIdToUpdate (since we're obviously updating it!)
-                // and also anything that depended on it.
-                if (projectIdToUpdate == projectId || existingTransitiveReferences?.Contains(projectId) == true)
-                {
-                    // This needs an update. If we know what to include in, we'll union it with the existing ones. Otherwise, we don't know
-                    // and we'll remove any data from the cache.
-                    if (newTransitiveReferences != null && existingTransitiveReferences != null)
-                    {
-                        builder[projectIdToUpdate] = existingTransitiveReferences.Union(newTransitiveReferences);
-                    }
-                    else
-                    {
-                        // Either we don't know the full set of the new references being added, or don't know the existing set projectIdToUpdate.
-                        // In this case, just remove it
-                        builder.Remove(projectIdToUpdate);
-                    }
-                }
-            }
-
-            return builder.ToImmutable();
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_reverseTransitiveReferencesMap"/> for the addition of new projects.
-        /// </summary>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseTransitiveReferencesMapForAdditionalProjectReferences(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseTransitiveReferencesMap,
-            ProjectId projectId,
-            IReadOnlyList<ProjectId> referencedProjectIds)
-        {
-            // To update the reverse transitive map, we need to add the existing reverse transitive references of projectId to any of referencedProjectIds,
-            // and anything else with a reverse dependency on them. If we don't already know our reverse transitive references, then we'll have to instead remove
-            // the cache entries instead of update them. We'll fetch this from the map, and use "null" to indicate the "we don't know and should remove the cache entry"
-            // instead
-            existingReverseTransitiveReferencesMap.TryGetValue(projectId, out var newReverseTranstiveReferences);
-
-            if (newReverseTranstiveReferences != null)
-            {
-                newReverseTranstiveReferences = newReverseTranstiveReferences.Add(projectId);
-            }
-
-            // We'll now loop through each entry in our existing cache and compute updates. We'll accumulate them into this builder.
-            var builder = existingReverseTransitiveReferencesMap.ToBuilder();
-
-            foreach (var projectIdToUpdate in existingReverseTransitiveReferencesMap.Keys)
-            {
-                existingReverseTransitiveReferencesMap.TryGetValue(projectIdToUpdate, out var existingReverseTransitiveReferences);
-
-                // The projects who need to have their caches updated are projectIdToUpdate (since we're obviously updating it!)
-                // and also anything that depended on us.
-                if (referencedProjectIds.Contains(projectIdToUpdate) || existingReverseTransitiveReferences?.Overlaps(referencedProjectIds) == true)
-                {
-                    // This needs an update. If we know what to include in, we'll union it with the existing ones. Otherwise, we don't know
-                    // and we'll remove any data from the cache.
-                    if (newReverseTranstiveReferences != null && existingReverseTransitiveReferences != null)
-                    {
-                        builder[projectIdToUpdate] = existingReverseTransitiveReferences.Union(newReverseTranstiveReferences);
-                    }
-                    else
-                    {
-                        // Either we don't know the full set of the new references being added, or don't know the existing set projectIdToUpdate.
-                        // In this case, just remove it
-                        builder.Remove(projectIdToUpdate);
-                    }
-                }
-            }
-
-            return builder.ToImmutable();
-        }
-
-        /// <summary>
-        /// Computes a new <see cref="_reverseTransitiveReferencesMap"/> for the removal of a project.
-        /// </summary>
-        /// <seealso cref="ComputeNewTransitiveReferencesMapForRemovedProject"/>
-        private static ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseTransitiveReferencesMapForRemovedProject(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseTransitiveReferencesMap,
-            ProjectId projectId)
-        {
-            var builder = existingReverseTransitiveReferencesMap.ToBuilder();
-
-            foreach (var (project, references) in existingReverseTransitiveReferencesMap)
-            {
-                if (references.Contains(projectId))
-                {
-                    // Invalidate the cache for projects that reference the removed project
-                    builder.Remove(project);
-                }
-            }
-
-            builder.Remove(projectId);
-            return builder.ToImmutable();
-        }
-
-        private ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> ComputeNewReverseTransitiveReferencesMapForRemovedProjectReference(
-            ImmutableDictionary<ProjectId, ImmutableHashSet<ProjectId>> existingReverseTransitiveReferencesMap,
-            ProjectId projectId,
-            ProjectId referencedProjectId)
-        {
-            var builder = existingReverseTransitiveReferencesMap.ToBuilder();
-
-            // Invalidate the transitive reverse references from every project previously referenced by the original
-            // project (transitively), except for cases where the removed project reference could not impact the result.
-            foreach (var (project, references) in existingReverseTransitiveReferencesMap)
-            {
-                if (!references.Contains(referencedProjectId))
-                {
-                    // If projectId references project, it isn't through referencedProjectId so the change doesn't
-                    // impact the dependency graph.
-                    //
-                    // Suppose we start with the following graph, and are removing the project reference A->B:
-                    //
-                    //   A -> B -> C
-                    //   A -> D
-                    //
-                    // This case is not hit for project C. The reverse transitive references for C contains B, which is
-                    // the target project of the removed reference. We can see that project C is impacted by this
-                    // removal, and after the removal, project A will no longer be in the reverse transitive references
-                    // of C.
-                    //
-                    // This case is hit for project D. The reverse transitive references for D does not contain B, which
-                    // means project D cannot be "downstream" of the impact of removing the reference A->B. We can see
-                    // that project A will still be in the reverse transitive references of D.
-                    //
-                    // This optimization does not catch all cases. For example, consider the following graph where we
-                    // are removing the project reference A->B:
-                    //
-                    //   A -> B -> D
-                    //   A -> D
-                    //
-                    // For this example, we do not hit this optimization because D contains project B in the set of
-                    // reverse transitive references. Without more complicated checks, we cannot rule out the
-                    // possibility that A may have been removed from the reverse transitive references of D by the
-                    // removal of the A->B reference.
-                    continue;
-                }
-
-                if (references.Contains(projectId))
-                {
-                    builder.Remove(project);
-                }
-            }
-
-            // Invalidate the transitive references from the previously-referenced project
-            builder.Remove(referencedProjectId);
-
-            return builder.ToImmutable();
-        }
-
-        internal ProjectDependencyGraph WithProjectRemoved(ProjectId projectId)
-        {
-            Contract.ThrowIfFalse(_projectIds.Contains(projectId));
-
-            // Project ID set and direct forward references are trivially updated by removing the key corresponding to
-            // the project getting removed.
-            var projectIds = _projectIds.Remove(projectId);
-            var referencesMap = ComputeNewReferencesMapForRemovedProject(_lazyReverseReferencesMap, _referencesMap, projectId);
-
-            // The direct reverse references map is updated by removing the key for the project getting removed, and
-            // also updating any direct references to the removed project.
-            var reverseReferencesMap = _lazyReverseReferencesMap is object
-                ? ComputeNewReverseReferencesMapForRemovedProject(_referencesMap, _lazyReverseReferencesMap, projectId)
-                : null;
-            var transitiveReferencesMap = ComputeNewTransitiveReferencesMapForRemovedProject(_transitiveReferencesMap, projectId);
-            var reverseTransitiveReferencesMap = ComputeNewReverseTransitiveReferencesMapForRemovedProject(_reverseTransitiveReferencesMap, projectId);
-            return new ProjectDependencyGraph(
-                projectIds,
-                referencesMap,
-                reverseReferencesMap,
-                transitiveReferencesMap,
-                reverseTransitiveReferencesMap,
-                topologicallySortedProjects: default,
-                dependencySets: default);
-        }
-
-        internal ProjectDependencyGraph WithProjectReferenceRemoved(ProjectId projectId, ProjectId referencedProjectId)
-        {
-            Contract.ThrowIfFalse(_projectIds.Contains(projectId));
-            Contract.ThrowIfFalse(_referencesMap[projectId].Contains(referencedProjectId));
-
-            // Removing a project reference doesn't change the set of projects
-            var projectIds = _projectIds;
-
-            // Incrementally update the graph
-            var referencesMap = _referencesMap.SetItem(projectId, _referencesMap[projectId].Remove(referencedProjectId));
-            var reverseReferencesMap = _lazyReverseReferencesMap is object
-                ? ComputeNewReverseReferencesMapForRemovedProjectReference(_lazyReverseReferencesMap, projectId, referencedProjectId)
-                : null;
-            var transitiveReferencesMap = ComputeNewTransitiveReferencesMapForRemovedProjectReference(_transitiveReferencesMap, projectId, referencedProjectId);
-            var reverseTransitiveReferencesMap = ComputeNewReverseTransitiveReferencesMapForRemovedProjectReference(_reverseTransitiveReferencesMap, projectId, referencedProjectId);
-
-            return new ProjectDependencyGraph(
-                projectIds,
-                referencesMap,
-                reverseReferencesMap,
-                transitiveReferencesMap,
-                reverseTransitiveReferencesMap,
-                topologicallySortedProjects: default,
-                dependencySets: default);
         }
 
         internal ProjectDependencyGraph WithProjectReferences(ProjectId projectId, IEnumerable<ProjectId> referencedProjectIds)
